# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OpenAI-совместимый HTTP API (chat, embeddings, models) и RAG поверх PostgreSQL/pgvector.",
        "title": "llama-cpp-gpt-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/rag/documents": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Список документов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResRAGDocuments"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            },
            "post": {
                "description": "Индексирует content или файл по path на сервере.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Индексация документа (JSON)",
                "parameters": [
                    {
                        "description": "Документ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ReqRAGIngest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResRAGIngest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/rag/documents/scan": {
            "post": {
                "description": "Индексирует файлы из каталогов RAG.AutoIngest в конфиге.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Сканирование каталогов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResRAGScan"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/rag/documents/upload": {
            "post": {
                "description": "Multipart upload: PDF, DOCX, TXT и др. — конвертация и индексация.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Загрузка файла",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Файл",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название документа",
                        "name": "title",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResRAGIngest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/rag/documents/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Документ по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID документа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.Document"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "rag"
                ],
                "summary": "Удалить документ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID документа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/rag/query": {
            "post": {
                "description": "Гибридный поиск (вектор + полнотекстовый) по проиндексированным документам.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Поиск по RAG",
                "parameters": [
                    {
                        "description": "Запрос",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ReqRAGQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResRAGQuery"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/rag/reindex-embeddings": {
            "post": {
                "description": "Пересчитывает embedding для всех чанков (после смены префиксов search_document/search_query).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Пересчёт эмбеддингов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ResRAGReindex"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.OpenAIErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "store.ChunkResult": {
            "type": "object",
            "properties": {
                "chunk_id": {
                    "type": "string"
                },
                "chunk_index": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object"
                },
                "score": {
                    "type": "number"
                },
                "source_path": {
                    "type": "string"
                },
                "text_rank": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "vector_rank": {
                    "type": "integer"
                }
            }
        },
        "store.Document": {
            "type": "object",
            "properties": {
                "chunk_count": {
                    "type": "integer"
                },
                "content_hash": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object"
                },
                "source_path": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "types.OpenAIErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/types.OpenAIErrorDetail"
                }
            }
        },
        "types.OpenAIErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "param": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "types.ReqRAGIngest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object"
                },
                "path": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "types.ReqRAGQuery": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "query": {
                    "type": "string"
                },
                "top_k": {
                    "type": "integer"
                }
            }
        },
...
        "types.ResRAGDocuments": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.Document"
                    }
                }
            }
        },
        "types.ResRAGIngest": {
            "type": "object",
            "properties": {
                "document": {
                    "$ref": "#/definitions/store.Document"
                }
            }
        },
        "types.ResRAGQuery": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.ChunkResult"
                    }
                }
            }
        },
        "types.ResRAGReindex": {
            "type": "object",
            "properties": {
                "reindexed_chunks": {
                    "type": "integer"
                }
            }
        },
        "types.ResRAGScan": {
            "type": "object",
            "properties": {
                "ingested": {
                    "type": "integer"
                }
            }
        },
...
    }
}